move props to component
because next does not allow to exporting if it is not a react component
