--- pages/[postId]/postId.props.ts
@@ -1,6 +1,0 @@
-import { PostProps, PageProps } from "../../utils/props";
-
-export interface BlogPostProps {
-    post: PostProps,
-    page: PageProps
-}
