--- pages/index/index.props.ts
@@ -1,7 +1,0 @@
-import { AuthStoreProps } from "../../stores/AuthStore/AuthStore";
-import { PostProps } from "../../utils/props";
-
-export interface Props {
-    authStore: AuthStoreProps
-    posts: Array<PostProps>
-}
